component: TextInput moved inside hexagon
