+++ src/components/DatabaseTables/StudentRecords/CoursesTable/components/EditCourseView.tsx
@@ -1,8 +1,8 @@
 import type { Course } from 'src/types/CoachingTypes';
 import type { EditableCourse, NewCourse } from '../types';
-import { Dropdown } from '@interface/components/FormComponents';
+import { Dropdown, TextInput } from '@interface/components/FormComponents';
 import React, { useState } from 'react';
-import { FormControls, TextInput } from 'src/components/FormComponents';
+import { FormControls } from 'src/components/FormComponents';
 import verifyRequiredInputs from 'src/components/FormComponents/functions/inputValidation';
 import ContextualView from 'src/hexagon/interface/components/Contextual/ContextualView';
 import { useContextualMenu } from 'src/hexagon/interface/hooks/useContextualMenu';
