+++ src/components/DatabaseTables/StudentRecords/CoursesTable/components/FilterCoursesTable.tsx
@@ -1,7 +1,7 @@
 import type { Dispatch, SetStateAction } from 'react';
 import type { FilterConfig } from 'src/components/Table/types';
+import { TextInput } from '@interface/components/FormComponents';
 import React from 'react';
-import { TextInput } from 'src/components/FormComponents';
 import { useContextualMenu } from 'src/hexagon/interface/hooks/useContextualMenu';
 
 export default function FilterCoursesTable({
