+++ src/components/DatabaseTables/StudentRecords/LessonsTable/components/FilterLessonsTable.tsx
@@ -1,6 +1,6 @@
 import type { Dispatch, SetStateAction } from 'react';
 import type { FilterConfig } from 'src/components/Table/types';
-import { TextInput } from 'src/components/FormComponents';
+import { TextInput } from '@interface/components/FormComponents';
 import { useContextualMenu } from 'src/hexagon/interface/hooks/useContextualMenu';
 
 interface FilterLessonsTableProps {
