+++ src/components/DatabaseTables/VocabQuizDb/QuizzesTable/components/FilterQuizTable.tsx
@@ -1,7 +1,7 @@
 import type { Dispatch, SetStateAction } from 'react';
 import type { FilterConfig } from 'src/components/Table/types';
+import { TextInput } from '@interface/components/FormComponents';
 import React from 'react';
-import { TextInput } from 'src/components/FormComponents/TextInput';
 import { useContextualMenu } from 'src/hexagon/interface/hooks/useContextualMenu';
 
 interface FilterQuizTableProps {
