+++ src/components/StudentDrillDown/components/StudentInfoCard.tsx
@@ -1,13 +1,12 @@
 import type { Coach, Student } from 'src/types/CoachingTypes';
 import CoachStudentDrillDown from '@interface/components/CoachStudentDrillDown/CoachStudentDrillDown';
-import { Dropdown } from '@interface/components/FormComponents';
+import { Dropdown, TextInput } from '@interface/components/FormComponents';
 import React, { useEffect, useState } from 'react';
 import pencil from 'src/assets/icons/pencil.svg';
 import {
   Checkbox,
   CoachDropdown,
   FormControls,
-  TextInput,
 } from 'src/components/FormComponents';
 import { toISODate } from 'src/hexagon/domain/functions/dateUtils';
 import ContextualView from 'src/hexagon/interface/components/Contextual/ContextualView';
