+++ src/hexagon/interface/components/FormComponents/index.ts
@@ -1,4 +1,7 @@
 import type { GenericDropdownOption } from '@interface/components/FormComponents/Dropdown/GenericDropdown';
 import Dropdown from '@interface/components/FormComponents/Dropdown/Dropdown';
 import GenericDropdown from '@interface/components/FormComponents/Dropdown/GenericDropdown';
-export { Dropdown, GenericDropdown, type GenericDropdownOption };
+import TextInput from '@interface/components/FormComponents/TextInput/TextInput';
+
+export { Dropdown, GenericDropdown, TextInput };
+export type { GenericDropdownOption };
